code refactored
changed the data type of discountPercentage from Integer to double

+++ pojo/src/main/java/com/stanzaliving/website/response/dto/ApartmentListingCardDto.java
@@ -60,7 +60,7 @@ public class ApartmentListingCardDto implements Serializable {
 	
 	private int startingPrice;
 	
-	private Integer discountPercentage;
+	private double discountPercentage;
 	
 	private Integer discountedPrice;
