+++ pojo/src/main/java/com/stanzaliving/website/response/dto/ResidenceResponseShortDTO.java
@@ -56,7 +56,7 @@ public class ResidenceResponseShortDTO implements Serializable {
 
 	private Integer unlockDiscountAmount;
 
-	private Integer discountPercentage;
+	private double discountPercentage;
 
 	private Integer discountedPrice;
